third_party/WebKit/LayoutTests/fast/animation/request-animation-frame-missing-arguments-expected.txt

Test how animation methods react to too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS requestAnimationFrame() threw exception TypeError: Failed to execute 'requestAnimationFrame' on 'DOMWindow': 1 argument required, but only 0 present..
PASS cancelAnimationFrame() threw exception TypeError: Failed to execute 'cancelAnimationFrame' on 'DOMWindow': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
